Move and remove files
