+++ src/cache_padded.rs
